+++ config-overrides.js
@@ -5,7 +5,8 @@ module.exports = function override(config) {
     alias: {
       ...config.alias,
       'components': path.resolve(__dirname, 'src/components'),
-      'redux2': path.resolve(__dirname, 'src/redux2'),
+      'containers': path.resolve(__dirname, 'src/containers'),
+      'store': path.resolve(__dirname, 'src/store'),
     },
   };
 return config;
