+++ src/index.tsx
@@ -6,7 +6,7 @@ import reportWebVitals from './reportWebVitals';
 import {Provider} from "react-redux"
 import thunk from "redux-thunk"
 import {createStore, applyMiddleware} from "redux"
-import reducer from "./redux2";
+import reducer from "store/reducer";
 
 const store = createStore(reducer, applyMiddleware(thunk))
