+++ src/store/reducer/DoingReducer.ts
@@ -1,27 +1,27 @@
-import {POST_TODO_REQ, POST_TODO_REQ_SUCCESS, POST_TODO_REQ_ERR, POST_TODO_REQ_RESET} from "./actions/types"
-const initialState={
+import {DOING_REQ, DOING_REQ_SUCCESS, DOING_REQ_ERR} from "store/types"
+import { DOING_ITEMS_DATA} from "store/actions/index"
+import {TodoItem} from "components/models"
+
+const initialState = {
     loading: false,
     success: false,
     error: false,
-    data: {}
+    data: Array<TodoItem>()
 }
 
 
 const reducer = (state = initialState, action:any) => {
     const newState = { ...state };
     switch (action.type) {
-      case POST_TODO_REQ_RESET:
-        return initialState
-        break;
-      case POST_TODO_REQ:
+      case DOING_REQ:
         newState.loading = true;
         break;
-      case POST_TODO_REQ_SUCCESS:
+      case DOING_REQ_SUCCESS:
         newState.loading = false;
         newState.success = true;
-        newState.data = action.payload
+        newState.data = [...DOING_ITEMS_DATA]
         break;
-      case POST_TODO_REQ_ERR:
+      case DOING_REQ_ERR:
         newState.loading = !state.loading;
         newState.error = true
     }
