+++ src/store/reducer/TodoReducer.ts
@@ -1,8 +1,7 @@
-import {DOING_REQ, DOING_REQ_SUCCESS, DOING_REQ_ERR} from "./actions/types"
-import { DOING_ITEMS_DATA} from "./actions/index"
-import {TodoItem} from "../components/models"
-
-const initialState = {
+import {TODO_REQ, TODO_REQ_SUCCESS, TODO_REQ_ERR} from "store/types"
+import { TODO_ITEMS_DATA} from "store/actions/index"
+import { TodoItem} from "components/models"
+const initialState={
     loading: false,
     success: false,
     error: false,
@@ -13,15 +12,15 @@ const initialState = {
 const reducer = (state = initialState, action:any) => {
     const newState = { ...state };
     switch (action.type) {
-      case DOING_REQ:
+      case TODO_REQ:
         newState.loading = true;
         break;
-      case DOING_REQ_SUCCESS:
+      case TODO_REQ_SUCCESS:
         newState.loading = false;
         newState.success = true;
-        newState.data = [...DOING_ITEMS_DATA]
+        newState.data = [...TODO_ITEMS_DATA]
         break;
-      case DOING_REQ_ERR:
+      case TODO_REQ_ERR:
         newState.loading = !state.loading;
         newState.error = true
     }
