+++ src/store/reducer/postTodoReducer.ts
@@ -1,26 +1,27 @@
-import {TODO_REQ, TODO_REQ_SUCCESS, TODO_REQ_ERR} from "./actions/types"
-import { TODO_ITEMS_DATA} from "./actions/index"
-import { TodoItem} from "../components/models"
+import {POST_TODO_REQ, POST_TODO_REQ_SUCCESS, POST_TODO_REQ_ERR, POST_TODO_REQ_RESET} from "store/types"
 const initialState={
     loading: false,
     success: false,
     error: false,
-    data: Array<TodoItem>()
+    data: {}
 }
 
 
 const reducer = (state = initialState, action:any) => {
     const newState = { ...state };
     switch (action.type) {
-      case TODO_REQ:
+      case POST_TODO_REQ_RESET:
+        return initialState
+        break;
+      case POST_TODO_REQ:
         newState.loading = true;
         break;
-      case TODO_REQ_SUCCESS:
+      case POST_TODO_REQ_SUCCESS:
         newState.loading = false;
         newState.success = true;
-        newState.data = [...TODO_ITEMS_DATA]
+        newState.data = action.payload
         break;
-      case TODO_REQ_ERR:
+      case POST_TODO_REQ_ERR:
         newState.loading = !state.loading;
         newState.error = true
     }
